services/python-worker/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
python-json-logger==2.0.7
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-kafka-python==0.42b0
opentelemetry-exporter-otlp==1.21.0
prometheus-client==0.19.0
